+++ src/cmds/stat.rs
@@ -1,3 +1,4 @@
+
 //! status command
 use super::Command;
 use colored::Colorize;